Basic Employee Get and GetAll Working, No Authorization Yet

+++ Controllers/UsersController.cs
@@ -32,22 +32,62 @@ namespace WebApi.Controllers
             _configuration = configuration;
         }
 
+        /// <summary>
+        /// Gets a list of API Users with varying detail depending on Role.
+        /// </summary>
+        /// <returns>A list of User Objects</returns>
+        /// <response code="200">A list of User Objects</response>
+        /// <response code="400">Token Provider User Doesn't exist any more</response>  
         // GET: api/Users
+        [ProducesResponseType(typeof(List<GetAllUser>), 200)]
         [HttpGet]
         public async Task<ActionResult<IEnumerable<GetAllUser>>> GetUser()
         {
-            return await _context.User
-                .Select(p => new GetAllUser
-                {
-                    Id = p.Id,
-                    FirstName = p.FirstName,
-                    LastName = p.LastName,
-                    Username = p.UserName,
-                    Role = p.Role
-                }).ToListAsync();
+            var identity = HttpContext.User.Identity as ClaimsIdentity;
+
+            IEnumerable<Claim> claim = identity.Claims;
+
+            var RoleClaim = claim
+                .Where(x => x.Type == ClaimTypes.Role)
+                .FirstOrDefault();
+
+            //if admin expose Usernames
+            if (RoleClaim.Value == "admin")
+            {
+                return await _context.User
+                   .Select(p => new GetAllUser
+                   {
+                       Id = p.Id,
+                       FirstName = p.FirstName,
+                       LastName = p.LastName,
+                       Username = p.UserName,
+                       Role = p.Role
+                   }).ToListAsync();
+            }
+            else //do not expose usernames
+            {
+                return await _context.User
+                   .Select(p => new GetAllUser
+                   {
+                       Id = p.Id,
+                       FirstName = p.FirstName,
+                       LastName = p.LastName,
+                       Role = p.Role
+                   }).ToListAsync();
+            }
         }
 
+        /// <summary>
+        /// Gets an API User Details.
+        /// </summary>
+        /// <param name="id" example="10"></param>
+        /// <returns>A newly created TodoItem</returns>
+        /// <response code="200">Returns the user item</response>
+        /// <response code="400">Token Provider User Doesn't exist any more</response>  
+        /// <response code="401">If the User is Unauthorized</response>  
+        /// <response code="404">If Username Not found</response>  
         // GET: api/Users/5
+        [ProducesResponseType(typeof(BasicUserResponse), 200)]
         [HttpGet("{id}")]
         public async Task<ActionResult<User>> GetUser(int id)
         {
@@ -62,20 +102,24 @@ namespace WebApi.Controllers
                 .Where(x => x.Type == ClaimTypes.Name)
                 .FirstOrDefault();
 
+
+
             // Finds user. using ID converted from string to int
             var userName = await _context.User
             .Where(x => x.Id == Int32.Parse(usernameClaim.Value))
             .FirstOrDefaultAsync();
 
-            //Username Token == null, then Badrequest
+
             if (userName == null)
             {
                 return BadRequest(new { message = "Woops" });
-            }
+            } 
+
             if (!(userName.Id == id) & userName.Role != "admin")
             {
-                return Unauthorized(new { message = "Trying to update other users if you're not an admin is not allowed. Message Administrator" });
+                return Unauthorized(new { message = "Trying to get other users if you're not an admin is not allowed. Message Administrator" });
             }
+
             var user = await _context.User.FindAsync(id);
 
             if (user == null)
@@ -83,8 +127,8 @@ namespace WebApi.Controllers
                 return NotFound();
             }
             //Password, hash and salt should never be exposed, I don't think admins should ever need to see a users password
-            //if needed they can change the password with an update/put request though.
-            return Ok(new
+            //if needed they can change the password with an update request though.
+            return Ok(new BasicUserResponse
             {
                 Id = user.Id,
                 Username = user.UserName,
@@ -94,7 +138,18 @@ namespace WebApi.Controllers
             }); ;
         }
 
-        // PUT: api/Users/5
+        /// <summary>
+        /// Updates an API Users Details.
+        /// </summary>
+        /// <param name="id" example="10"></param>
+        /// <param name="user"></param>
+        /// <returns>A message saying user was updated.</returns>
+        /// <response code="200">Returns the Updated user Id and Name</response>
+        /// <response code="400">Token Provider User Doesn't exist any more</response>  
+        /// <response code="401">If the User is Unauthorized</response>  
+        /// <response code="404">If Username Not found</response>  
+        /// 
+        // PUT: api/Users/5 (UPDATE)
         [HttpPut("{id}")]
         public async Task<IActionResult> PutUser(int id, UpdateUser user)
         {
@@ -112,7 +167,9 @@ namespace WebApi.Controllers
             // Finds user. using ID converted from string to int
             var userName = await _context.User
             .Where(x => x.Id == Int32.Parse(usernameClaim.Value))
+            .AsNoTracking()
             .FirstOrDefaultAsync();
+
 
 
 
@@ -122,34 +179,34 @@ namespace WebApi.Controllers
                 return BadRequest(new { message = "Woops" });
             }
 
-            if ((id < 0) & (user.Id == null))
-            {
+            bool isadmin = (userName.Role != "admin");
 
+            if ((id < 0))
+            {
                 return BadRequest();
             }
 
             // if Username.Id (The id of the token != id of the user being edited, unauthorized;
             //and user is not an admin return unauthorized.
-            if (!(userName.Id == user.Id.GetValueOrDefault()) & userName.Role != "admin")
+            if (!(userName.Id == id) & isadmin)
             {
                 return Unauthorized(new { message = "Trying to update other users if you're not an admin is not allowed. Message Administrator" });
+            }
+            //throw unauthorized, if the user is not an admin and is trying to set a role !
+            if (!(user.Role == null) & isadmin)
+            {
+                return Unauthorized(new { message = "User can't adjust role, please try a different update command without including role." });
             }
 
             var userobj = await _context.User
             .FindAsync(id);
 
             //only admin can adjust role
-            if (userobj.Role == "admin" && user.Role != null)
+            if (isadmin & user.Role != null)
             {
                 userobj.Role = user.Role;
             }
 
-            //throw unauthorized, if the user is not an admin and is trying to set a role !
-            if (userobj.Role != "admin" && user.Role != null)
-            {
-
-                return Unauthorized(new { message = "User can't adjust role, please try a different update command without including role."});
-            }
 
             //if password change is being requested, we need to update the password, hash & salt.
             if(user.Password != null)
@@ -199,7 +256,7 @@ namespace WebApi.Controllers
                 }
             }
 
-            return Ok(new{ message = "Updated User: " + userName.Id + " " + userName.UserName });
+            return Ok(new{ message = "Updated User: " + userobj.Id + " " + userobj.UserName });
 
         }
 
@@ -211,7 +268,6 @@ namespace WebApi.Controllers
         /// <response code="201">Returns the newly created user item</response>
         /// <response code="400">If the item is null, password doesn't exist or username already exists</response>  
         // POST: api/Users
-    [HttpPost]
         [ProducesResponseType(StatusCodes.Status201Created)]
         [ProducesResponseType(StatusCodes.Status400BadRequest)]
         [HttpPost]
@@ -293,7 +349,8 @@ namespace WebApi.Controllers
             {
                 Subject = new ClaimsIdentity(new Claim[]
                 {
-                    new Claim(ClaimTypes.Name, userobj.Id.ToString())
+                    new Claim(ClaimTypes.Name, userobj.Id.ToString()),
+                    new Claim(ClaimTypes.Role, userobj.Role.ToString())
                 }),
                 Expires = DateTime.UtcNow.AddDays(7),
                 SigningCredentials = new SigningCredentials(new SymmetricSecurityKey(key), SecurityAlgorithms.HmacSha256Signature)
@@ -314,21 +371,53 @@ namespace WebApi.Controllers
         }
 
 
-
+        /// <summary>
+        /// Deletes an API User.
+        /// </summary>
+        /// <param name="id"></param>
+        /// <returns>return deleted user info</returns>
+        /// <response code="200">return deleted user info</response>
+        /// <response code="403">Throws forbidden if role user tries to delete anything</response>  
         // DELETE: api/Users/5
         [HttpDelete("{id}")]
         public async Task<ActionResult<User>> DeleteUser(int id)
         {
-            var user = await _context.User.FindAsync(id);
-            if (user == null)
+            var identity = HttpContext.User.Identity as ClaimsIdentity;
+
+            IEnumerable<Claim> claim = identity.Claims;
+
+            var RoleClaim = claim
+                .Where(x => x.Type == ClaimTypes.Role)
+                .FirstOrDefault();
+            //if admin Delete User
+            if (RoleClaim.Value == "admin")
             {
-                return NotFound();
+                var user = await _context.User.FindAsync(id);
+                if (user == null)
+                {
+                    return NotFound();
+                }
+
+                _context.User.Remove(user);
+                await _context.SaveChangesAsync();
+
+                return Ok(new BasicUserResponse
+                {
+                    Id = user.Id,
+                    Username = user.UserName,
+                    FirstName = user.FirstName,
+                    LastName = user.LastName,
+                    Role = user.Role,
+                }); ;
+
+            }
+            else //if not admin throw forbidden
+            {
+                return Forbid();
             }
 
-            _context.User.Remove(user);
-            await _context.SaveChangesAsync();
 
-            return user;
+
         }
 
         private bool UserExists(int id)
